+++ gear/form_transformer/src/python/form_csv_app/run.py
@@ -3,19 +3,22 @@
 import logging
 from typing import Optional
 
+from datastore.forms_store import FormsStore
 from flywheel.rest import ApiException
-from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
+from flywheel_adaptor.flywheel_proxy import ProjectAdaptor, ProjectError
 from flywheel_gear_toolkit.context.context import GearToolkitContext
 from gear_execution.gear_execution import (
     ClientWrapper,
-    ContextClient,
+    GearBotClient,
     GearEngine,
     GearExecutionEnvironment,
     GearExecutionError,
     InputFileWrapper,
 )
 from inputs.parameter_store import ParameterStore
+from keys.keys import DefaultValues
 from outputs.errors import ListErrorWriter
+from preprocess.preprocessor import FormPreprocessor, FormProjectConfigs
 from pydantic import ValidationError
 from transform.transformer import FieldTransformations, TransformerFactory
 from utils.utils import parse_string_to_list
@@ -29,9 +32,11 @@ class FormCSVtoJSONTransformer(GearExecutionEnvironment):
     """Visitor for the templating gear."""
 
     def __init__(self, client: ClientWrapper, file_input: InputFileWrapper,
+                 config_input: InputFileWrapper,
                  transform_input: Optional[InputFileWrapper]) -> None:
         self.__client = client
         self.__file_input = file_input
+        self.__config_input = config_input
         self.__transform_input = transform_input
 
     @classmethod
@@ -53,17 +58,25 @@ class FormCSVtoJSONTransformer(GearExecutionEnvironment):
           GearExecutionError if any expected inputs are missing
         """
 
-        client = ContextClient.create(context=context)
+        assert parameter_store, "Parameter store expected"
+
+        client = GearBotClient.create(context=context,
+                                      parameter_store=parameter_store)
 
         file_input = InputFileWrapper.create(input_name='input_file',
                                              context=context)
-        assert file_input, "create raises exception if missing expected input"
+        assert file_input, "missing expected input, input_file"
+
+        form_configs_input = InputFileWrapper.create(
+            input_name='form_configs_file', context=context)
+        assert form_configs_input, "missing expected input, form_configs_file"
 
         transform_input = InputFileWrapper.create(input_name='transform_file',
                                                   context=context)
 
         return FormCSVtoJSONTransformer(client=client,
                                         file_input=file_input,
+                                        config_input=form_configs_input,
                                         transform_input=transform_input)
 
     def run(self, context: GearToolkitContext) -> None:
@@ -91,15 +104,23 @@ class FormCSVtoJSONTransformer(GearExecutionEnvironment):
         downstream_gears = parse_string_to_list(
             context.config.get('downstream_gears', None))
 
+        prj_adaptor = ProjectAdaptor(project=project, proxy=proxy)
+
+        error_writer = ListErrorWriter(container_id=file_id,
+                                       fw_path=proxy.get_lookup_path(file))
+
+        preprocessor = self.__get_preprocessor(
+            form_config_input=self.__config_input,
+            ingest_project=prj_adaptor,
+            error_writer=error_writer)
+
         with open(self.__file_input.filepath, mode='r',
                   encoding='utf-8') as csv_file:
-            error_writer = ListErrorWriter(container_id=file_id,
-                                           fw_path=proxy.get_lookup_path(file))
             success = run(input_file=csv_file,
-                          destination=ProjectAdaptor(project=project,
-                                                     proxy=proxy),
+                          destination=prj_adaptor,
                           transformer_factory=self.__build_transformer(
                               self.__transform_input),
+                          preprocessor=preprocessor,
                           error_writer=error_writer,
                           gear_name=gear_name,
                           downstream_gears=downstream_gears)
@@ -136,14 +157,58 @@ class FormCSVtoJSONTransformer(GearExecutionEnvironment):
                 return TransformerFactory(
                     FieldTransformations.model_validate_json(json_file.read()))
             except ValidationError as error:
-                raise GearExecutionError('Error reading transformation file'
-                                         f'{error}') from error
+                raise GearExecutionError(
+                    'Error reading transformation file'
+                    f'{transformer_input.filename}: {error}') from error
+
+    def __get_preprocessor(self, form_config_input: InputFileWrapper,
+                           ingest_project: ProjectAdaptor,
+                           error_writer: ListErrorWriter) -> FormPreprocessor:
+        """Reads the forms config file and initialize the preprocessor.
+
+        Args:
+          form_config_input: the input file wrapper for form configs file
+
+        Returns:
+          the FormPreprocessor with given configs
+        """
+
+        with open(form_config_input.filepath, mode='r',
+                  encoding='utf-8') as configs_file:
+            try:
+                form_configs = FormProjectConfigs.model_validate_json(
+                    configs_file.read())
+            except ValidationError as error:
+                raise GearExecutionError(
+                    'Error reading form configurations file'
+                    f'{form_config_input.filename}: {error}') from error
+
+            legacy_label = (form_configs.legacy_project_label
+                            if form_configs.legacy_project_label else
+                            DefaultValues.LEGACY_PRJ_LABEL)
+            try:
+                legacy_project = ProjectAdaptor.create(
+                    proxy=ingest_project.proxy,
+                    group_id=ingest_project.group,
+                    project_label=legacy_label)
+            except ProjectError as error:
+                raise GearExecutionError(
+                    f'Failed to retrieve legacy project: {error}') from error
+
+            forms_store = FormsStore(ingest_project=ingest_project,
+                                     legacy_project=legacy_project)
+
+            return FormPreprocessor(primary_key=form_configs.primary_key,
+                                    forms_store=forms_store,
+                                    module_info=form_configs.module_configs,
+                                    error_writer=error_writer)
 
 
 def main():
     """Main method for Form CSV to JSON Transformer."""
 
-    GearEngine().run(gear_type=FormCSVtoJSONTransformer)
+    GearEngine.create_with_parameter_store().run(
+        gear_type=FormCSVtoJSONTransformer)
 
 
 if __name__ == "__main__":