+++ docs/form_transformer/CHANGELOG.md
@@ -2,17 +2,20 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.1.0
+* Adds pre-processing checks.
+  
 ## 1.0.5
-* Update error reporting - move error metadata to visit error log files stored at project level.
+* Updates error reporting - move error metadata to visit error log files stored at project level.
   
 ## 1.0.2
 - Removes GearBot client
-- Change acquisition file naming schema
+- Changes acquisition file naming schema
 
 ## 1.0.0
 
 - Moves form-specific functionality of csv-to-json-transformer to form-transformer gear.
-- Add transformer schema as input file
+- Adds transformer schema as input file
 
 ## 0.0.11 (from CSV-to-JSON-transformer)
 - Normalizes the visit date to `YYYY-MM-DD` format